SQL/Eniac project - project 1
We have done Eniac project in first week of my data science bootcamp. Eniac is spanish compnay, who is selling high-tech products that they wanted to expand in Brazil. Then Eniac contacted Magist. Magist is a Brazilian Software as a service company which offers to connect small and medium-sized stores with the Biggest Brazilian Marketplace. Here we explored Magist database.